Escape reserved characters when generating urls.
Uses rfc3986 reserved characters, without /, because slashes in the
s3 object name have already been accounted for when the name was created.
(i.e. the end-to-end expectation is that they are used as namespace delimiters
when used in URLs).

This patch also escapes spaces with percent encoding.

More info on what is considered reserved:

rfc 3986 specifies these as reserved characters in a URL:

    !*'();:@&=+$,/?#[]

of these, Amazon thinks these need to be escaped when it generates urls
from its web console:

    ';:@&=+$,?#[ ]

and Amazon thinks these DO NOT need to be escaped:

    !*()

Amazon escapes spaces with +

Ruby's URI provides two regexes. What it considers to be reserved is an inexplicable
subset of rfc3986

    URI::REGEXP::PATTERN::UNRESERVED
    -_.!~*'()a-zA-Z\\d

    URI::REGEXP::PATTERN::RESERVED
    ;/?:@&=+$,\\[\\]