Use the options passed to S3Object.copy, supporting reduced redundancy
I need to enable reduced redundancy on some existing S3 objects.
The typical way to do this is copy the object within S3, supplying the
new storage-class (reduced redundancy) as an option to the copy.
However, aws-s3 is ignoring the copy options, instead only using the
default_options to specify the copy-source.

Update the copy method to merge the copy-source options with the passed
options, so that storage-class can be specified.